Implement all Rotation trait methods.
